ftp/requirements.txt
# requirements.txt

annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
Werkzeug==3.0.2
